requirements.txt
camelot-py[all]
ghostscript
opencv-python